Handle length modifiers l and h for d, i, u, o, x, X specifiers

+++ print_octal.c
@@ -1,19 +1,17 @@
 #include "main.h"
 
 /**
- * print_octal - prints an unsigned int in octal
- * @n: number to print
- *
- * Return: number of characters printed
+ * print_octal - print an unsigned long in octal
+ * @n: number
+ * Return: count of characters printed
  */
-int print_octal(unsigned int n)
+int print_octal(unsigned long n)
 {
-	int count = 0;
+    int count = 0;
 
-	if (n / 8)
-		count += print_octal(n / 8);
+    if (n / 8)
+        count += print_octal(n / 8);
 
-	count += _putchar((n % 8) + '0');
-
-	return (count);
+    count += _putchar((n % 8) + '0');
+    return (count);
 }
