+++ print_unsigned.c
@@ -1,19 +1,17 @@
 #include "main.h"
 
 /**
- * print_unsigned - prints an unsigned int in base 10
- * @n: number to print
- *
- * Return: number of characters printed
+ * print_unsigned - print an unsigned long integer
+ * @n: number
+ * Return: count of characters printed
  */
-int print_unsigned(unsigned int n)
+int print_unsigned(unsigned long n)
 {
-	int count = 0;
+    int count = 0;
 
-	if (n / 10)
-		count += print_unsigned(n / 10);
+    if (n / 10)
+        count += print_unsigned(n / 10);
 
-	count += _putchar((n % 10) + '0');
-
-	return (count);
+    count += _putchar((n % 10) + '0');
+    return (count);
 }
